requirements.txt
pytesseract==0.3.13
pdfplumber==0.11.4
Pillow==10.4.0
opencv-python==4.10.0.84
#pandas==2.2.2
python-dotenv==1.0.1
cryptography==43.0.1
selenium==4.21.0
beautifulsoup4==4.12.3
lxml==5.3.0
rapidfuzz==3.9.6
requests==2.32.3
loguru==0.7.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
